refactor: 커스텀 메뉴 분할
